+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+
+## 1.0.1
+
+* add interceptors in 'Rest' annotation
+* add the field 'name' which the annotations of method parameters
+* add the parameters of the 'Rest', resultData,resultMessage,resultCode
+* add header in 'Get'、'Post'、'Put'、'Delete' annotation
+* add @Header annotation as parameters 
+* fix bugs
+
+## 1.0.0
+
+* add  return 
+* fix bugs
+
+
 ## 0.0.1
 
 * TODO: Describe initial release.